evals/requirements.txt
# Additional requirements for OpenAI evals integration
pyyaml>=6.0
python-dotenv>=1.0.0
